Moved Activity into a section including sub-components
