--- source/components/Activity/activity.stories.jsx
@@ -1,11 +1,0 @@
-import Activity from './index';
-
-export default {
-  title: 'Components/Activity',
-  component: Activity,
-};
-
-const Template = (args) => <Activity {...args} />;
-
-export const Default = Template.bind({});
-Default.args = {};
