+++ source/components/index.js
@@ -6,7 +6,6 @@ export { default as NavBar } from './NavBar';
 export { default as Layout } from './Layout';
 export { default as Actions } from './Actions';
 export { default as Tokens } from './Tokens';
-export { default as Activity } from './Activity';
 export { default as Apps } from './Apps';
 export { default as RevealSeedPhrase } from './RevealSeedPhrase';
 export { default as SeedPhrase } from './SeedPhrase';
@@ -35,7 +34,6 @@ export { default as ActionButton } from './ActionButton';
 export { default as Tabs } from './Tabs';
 export { default as Tab } from './Tab';
 export { default as AssetItem } from './AssetItem';
-export { default as ActivityItem } from './ActivityItem';
 export { default as GenericIcon } from './GenericIcon';
 export { default as AppItem } from './AppItem';
 export { default as LinkButton } from './LinkButton';
