+++ source/views/Extension/Views/Home/sections/Activity/components/ActivityItem/components/ActivityItemDetails/index.js
