+++ source/views/Extension/Views/Home/sections/Activity/components/ActivityItem/components/ActivityItemDetails/styles.js
