+++ source/views/Extension/Views/Home/sections/Activity/components/ActivityItem/components/ActivityItemDisplay/index.js
