+++ source/views/Extension/Views/Home/sections/Activity/components/ActivityItem/components/items/Liquidity/components/LiquidityIcon/index.jsx
@@ -3,7 +3,7 @@ import clsx from 'clsx';
 import PropTypes from 'prop-types';
 import UnknownIcon from '@assets/icons/unknown-icon.svg';
 import { currencyPropTypes } from '@shared/constants/currencies';
-import TokenIcon from '../../../../../../TokenIcon';
+import { TokenIcon } from '@components';
 
 import useStyles from './styles';
