+++ source/views/Extension/Views/Home/sections/Activity/components/ActivityItem/components/items/Liquidity/components/LiquidityIcon/styles.js
