+++ source/views/Extension/Views/Home/sections/Activity/components/ActivityItem/components/items/NFT.js
@@ -1,37 +1,32 @@
 import React from 'react';
-import NumberFormat from 'react-number-format';
 import moment from 'moment';
 import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
+
 import { capitalize } from '@material-ui/core';
-
 import UnknownIcon from '@assets/icons/unknown-icon.svg';
+import shortAddress from '@shared/utils/short-address';
 
-import GenericIcon from '../../../GenericIcon';
+import { GenericIcon } from '@components';
+
 import ActivityItemDisplay from '../ActivityItemDisplay';
 import ActivityItemDetails from '../ActivityItemDetails';
-
 import { getSubtitle } from '../../utils';
 
-const TokenItem = (props) => {
-  const {
-    type,
-    to,
-    from,
-    amount,
-    value,
-    date,
-    symbol,
-    logo,
-    canisterId,
-    details,
-    setOpenDetail,
-    isTransaction,
-    hovering,
-    copied,
-    onCopy,
-    tooltipText,
-  } = props;
+const NFTItem = ({
+  type,
+  to,
+  from,
+  date,
+  logo,
+  canisterId,
+  details,
+  setOpenDetail,
+  hovering,
+  copied,
+  onCopy,
+  tooltipText,
+}) => {
   const { t } = useTranslation();
 
   return (
@@ -42,8 +37,8 @@ const TokenItem = (props) => {
             image={logo || UnknownIcon}
             type={type}
           />
-          )}
-        title={`${capitalize(type?.toLowerCase())} ${symbol ?? ''}`}
+        )}
+        title={`${capitalize(type?.toLowerCase())} NFT`}
         subtitle={moment(date).format('MMM Do')}
         tooltip={getSubtitle(type, to, from, t, canisterId)}
         copied={copied}
@@ -51,49 +46,41 @@ const TokenItem = (props) => {
         onCopy={onCopy}
       />
       <ActivityItemDetails
-        main={<NumberFormat value={amount} displayType="text" thousandSeparator="," suffix={` ${symbol}`} decimalScale={5} />}
-        secondary={<NumberFormat value={value} displayType="text" thousandSeparator="," prefix="$" suffix=" USD" decimalScale={2} />}
+        main={details?.tokenId?.length > 5 ? shortAddress(details?.tokenId) : `#${details?.tokenId}`}
+        secondary={details?.canisterInfo?.name || shortAddress(canisterId)}
         hovering={hovering}
         details={details}
         setOpenDetail={setOpenDetail}
-        isTransaction={isTransaction}
       />
     </>
   );
 };
 
-export default TokenItem;
+export default NFTItem;
 
-TokenItem.defaultProps = {
+NFTItem.defaultProps = {
   to: null,
   from: null,
-  amount: null,
-  value: null,
   type: 'PLUG',
-  hovering: false,
+  name: null,
   canisterId: null,
   details: null,
+  hovering: false,
 };
 
-TokenItem.propTypes = {
+NFTItem.propTypes = {
   type: PropTypes.number,
   canisterId: PropTypes.string,
   details: PropTypes.objectOf(PropTypes.string),
   to: PropTypes.string,
   from: PropTypes.string,
-  amount: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.number,
-  ]),
-  value: PropTypes.number,
-  symbol: PropTypes.string.isRequired,
   logo: PropTypes.string.isRequired,
   date: PropTypes.oneOfType([
     PropTypes.instanceOf(Date),
     PropTypes.string,
   ]).isRequired,
+  name: PropTypes.string,
   setOpenDetail: PropTypes.func.isRequired,
-  isTransaction: PropTypes.bool.isRequired,
   hovering: PropTypes.bool,
   copied: PropTypes.bool.isRequired,
   onCopy: PropTypes.func.isRequired,
