+++ source/views/Extension/Views/Home/sections/Activity/components/ActivityItem/components/items/Swap/components/SwapIcon/index.jsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import SwapArrowImg from '@assets/icons/swap-arrow.svg';
 import UnknownIcon from '@assets/icons/unknown-icon.svg';
 import { currencyPropTypes } from '@shared/constants/currencies';
-import TokenIcon from '../../../TokenIcon';
+import { TokenIcon } from '@components';
 
 import useStyles from './styles';
