+++ source/views/Extension/Views/Home/sections/Activity/components/ActivityItem/components/items/Swap/components/SwapIcon/styles.js
