+++ source/views/Extension/Views/Home/sections/Activity/components/ActivityItem/components/items/Swap/components/index.js
@@ -1,0 +1,1 @@
+export { default as SwapIcon } from './SwapIcon';
