+++ source/views/Extension/Views/Home/sections/Activity/components/ActivityItem/components/items/index.js
@@ -1,0 +1,6 @@
+export { default as LiquidityItem } from './Liquidity';
+export { default as SwapItem } from './Swap';
+export { default as JellyItem } from './Jelly';
+export { default as NFTItem } from './NFT';
+export { default as PlugItem } from './Plug';
+export { default as TokenItem } from './Token';
