+++ source/views/Extension/Views/Home/sections/Activity/components/ActivityItem/styles.js
