+++ source/views/Extension/Views/Home/sections/Activity/components/ActivityItem/utils.js
