+++ source/views/Extension/Views/Home/sections/Activity/components/index.js
@@ -1,0 +1,2 @@
+export { default as EmptyState } from './EmptyState';
+export { default as ActivityItem } from './ActivityItem';
