+++ source/views/Extension/Views/Home/sections/Activity/styles.js
