+++ source/views/Extension/Views/Home/sections/index.js
@@ -1,0 +1,2 @@
+export { default as Activity } from './Activity';
+
